# Conflicts:
#	Common/gfx/allegrobitmap.cpp
#	Common/gfx/allegrobitmap.h
#	Engine/ac/display.cpp
#	Engine/ac/draw.cpp
#	Engine/main/game_run.cpp
#	Engine/media/video/video.cpp
#	Engine/media/video/videoplayer.cpp
#	Engine/media/video/videoplayer.h
#	Engine/script/cc_instance.h
#	Engine/script/scriptexecutor.cpp
#	Solutions/Engine.App/Engine.App.vcxproj
#	Solutions/Engine.App/Engine.App.vcxproj.filters

+++ Engine/media/audio/soundclip.cpp
@@ -40,7 +40,7 @@ SoundClip::SoundClip(int slot, AudioFileType snd_type, bool loop)
 
     const auto player = audio_core_get_player(slot);
     lengthMs = (int)std::round(player->GetDurationMs());
-    freq = player->GetFrequency();
+    freq = static_cast<int>(player->GetFrequency());
 }
 
 SoundClip::~SoundClip()
